+++ taylor_series/sin.go
@@ -9,10 +9,10 @@ func Sin(x float64) float64 {
 	term := x
 	sin := x
 
-	for i := 3; term != 0.0; i += 2 {
+	for i := 3.0; term != 0.0; i += 2.0 {
 		// Sin(x) = x - x^3/3! + x^5/5! - x^7/7! + ...
 		// OR  pow(-1, n) * pow(x, 2n+1) / fac(2n+1)
-		term *= (-x * x) / float64(i*(i-1))
+		term *= (-x * x) / (i * (i - 1))
 		sin += term
 	}
 	return sin
